Minor clean-up, move RNodeBool to seperate file, and created RNodeDouble.
