+++ src/rnodes/mod.rs
@@ -1,2 +1,4 @@
 pub mod rnode;
+pub mod rnode_bool;
+pub mod rnode_double;
