+++ src/rnodes/rnode_double.rs
@@ -1,0 +1,23 @@
+use crate::rnodes::rnode::{EnumNodeType, RNode};
+
+pub struct RNodeDouble
+{
+    pub value: f64,
+}
+
+impl RNodeDouble
+{
+    pub fn new(value: f64) -> Self
+    {
+        Self { value }
+    }
+}
+
+impl RNode for RNodeDouble
+{
+    fn get_node_type(&self) -> EnumNodeType
+    {
+        return EnumNodeType::BOOL;
+    }
+}
+
